Update landing-page.js
Event Listeners are not getting recognised by Chrome DevTools.
My suspicion is that the "TypeError/property of null" comes from the fact that I'm assigning listeners to elements that haven't yet been rendered, so I have put them into "window.onload" to see if that will rectify the problem. The only problem with this theory is that the variables are set just fine.